requirements.txt
streamlit==1.29.0
openai==1.3.0
langchain==0.0.340
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
tiktoken==0.5.1
faiss-cpu==1.7.4
pypdf==3.17.1
